videojuegos/urls.py
from rest_framework import routers
from django.urls import path, include
from . import views

router = routers.DefaultRouter()
router.register("", views.VideoJuegosView)
urlpatterns = [
path('', include(router.urls)),
]