requirements.txt
Django==2.1.4
djangorestframework == 3.8.2